Function glm::builtin::packDouble2x32

pub fn packDouble2x32(v: UVec2) -> f64

Returns a double-precision value obtained by packing the components of v into a 64-bit value.

If an IEEE 754 Iinf or NaN is created, it will not signal, and the resulting floating point value is unspecified. Otherwise, the bit-level representation of v is preserved. The first vector component specifies the 32 least significant bits; the second component specifies the 32 most significant bits.

Example