Module glm::ext

Functions that extend the functionalities of GLSL data types.

Note

Unlike built-in functions, extension functions are not re-exported to the root module of glm. You have to use glm::ext.

Also unlike the built-in functions, the naming of extension functions follows the Rust convention, instead of the GLSL's. That is, snake case is used.

The parameters of matrix related functions are passed by reference.

Reexports

pub use self::consts::{Consts, epsilon, pi, tau, root_pi, half_pi, one_third_pi, quarter_pi, one_over_pi, one_over_tau, two_over_pi, four_over_pi, two_over_root_pi, one_over_root_two, root_half_pi, root_tau, root_ln_four, e, euler, root_two, root_three, root_five, ln_two, ln_ten, ln_ln_two, one_third, two_thirds, golden_ratio};

Modules

consts

Functions

angle
Returns angle between vectors x and y.

cbrt
Returns the cubic root.

is_invertible
Returns true if the square matrix m is invertible, i.e., its determinant does not close or equal to 0.

is_perpendicular
Returns true if vector x is perpendicular to y, i.e., angle between x and y is π/2.

look_at
Build a look at view matrix based on the default handedness.

look_at_rh
Build a right handed look at view matrix.

normalize_to
Normalizes vector x of specific length len.

perspective
Creates a matrix for a symetric perspective-view frustum based on the default handedness.

perspective_rh
Creates a matrix for a right handed, symetric perspective-view frustum.

pow2
x * x.

pow3
x * x * x.

powi
Raise a number to an integer power.

projection
Projects x on y.

recip
Returns the reciprocal (inverse) of float number x.

recip_length
Returns the reciprocal (inverse) of the length of vector x.

rotate
Builds a rotation 4 * 4 matrix created from an axis vector and an angle.

scale
Builds a scale 4 * 4 matrix created from 3 scalars.

sin_cos
Simultaneously computes the sine and cosine of x, returns (sin(x), cos(x)).

sqlength
Returns the squre of the length of vector x.

trace
Returns the trace of a square matrix m.

translate
Builds a translation 4 * 4 matrix created from a vector of 3 components.